+++ packages/ai_logger/lib/ai_logger.dart
@@ -5,3 +5,4 @@ export 'package:ai_logger_core/ai_logger_core.dart';
 export 'src/flutter_error_classifier.dart';
 export 'src/flutter_hooks.dart';
 export 'src/route_observer.dart';
+export 'src/web_runtime_error_classifier.dart';