+++ packages/ai_logger/lib/src/web_runtime_hooks_stub.dart
@@ -1,0 +1,5 @@
+import 'package:ai_logger_core/ai_logger_core.dart' as core;
+
+void installWebRuntimeHooks({core.AiLogger? target}) {}
+
+void resetWebRuntimeHooksForTesting() {}